[Agent] add IO_EVENT EbpfType, and filter it temporarily
