+++ agent/src/ebpf/kernel/include/socket_trace_common.h
@@ -142,6 +142,20 @@ struct allow_port_bitmap {
 	__u8 bitmap[65536 / 8];
 } __attribute__((packed));
 
+struct __io_event_buffer {
+	__u32 bytes_count;
+
+	// 0: write
+	// 1: read
+	__u32 operation;
+
+	// nanosecond
+	__u64 latency;
+
+	// strings terminated with \0
+	char filename[64];
+} __attribute__((packed));
+
 #define is_set_bitmap(bitmap, idx) (bitmap[(idx) / 8] & (1 << ((idx) % 8)))
 
 // struct ebpf_proc_info -> offsets[]  arrays index.
